int -> long

+++ trunk/source/Library-ModelCheckerAssistant/src/tw/ntu/svvrl/ultimate/lib/modelcheckerassistant/state/programstate/ForkHandler.java
@@ -24,7 +24,7 @@ public class ForkHandler {
 	 * 		and update the original thread state.)
 	 */
 	public ProgramState doFork() {
-		final int currentThreadID = mTrans.getThreadID();
+		final long currentThreadID = mTrans.getThreadID();
 		final ThreadState currentThreadState = mProgramState.getThreadStateByID(currentThreadID);
 		final ForkStatement forkStmt = ((ForkThreadCurrent) mTrans.getIcfgEdge()).getForkStatement();
 		final String forkProcName = forkStmt.getProcedureName();
